dlt/pyproject.toml
[project]
name = "beefy-databarn-dlt"
version = "0.1.0"
description = "Beefy Databarn - DLT data loading pipelines"
requires-python = ">=3.12,<3.14"
dependencies = [
    "apscheduler>=3.11.1",
    "dlt==1.18.2",
    "dlt[clickhouse,filesystem,sql_database]>=1.18.2",
    "httpx>=0.27.0",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
    "clickhouse-connect>=0.10.0",
    "psutil>=7.1.3",
    "psycopg2-binary>=2.9.0",
    "json5>=0.12.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "py-spy>=0.4.1", # profiling
]

# This is a data project, not a Python package - disable building
[tool.uv]
package = false

[dependency-groups]
dev = [
    "dlt[workspace]>=1.18.2",
]

